+++ exts/omni.warp/docs/CHANGELOG.md
@@ -21,6 +21,7 @@
   ([GH-963](https://github.com/NVIDIA/warp/issues/963)).
 - Fix scaling not being correctly applied to rendered meshes in some cases
   ([GH-880](https://github.com/NVIDIA/warp/issues/880)).
+- Fix handling of generic kernels with `wp.jax_experimental.ffi.jax_kernel()`.
 
 ## [1.9.0] - 2025-09-04